Move stuff in/to src/Elewant/Webapp/DomainModel
