+++ src/Elewant/Webapp/DomainModel/Contributor/EmptyContributorList.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Elewant\Webapp\DomainModel\Contributor;
+
+final class EmptyContributorList implements ContributorList
+{
+    public function allContributors(): array
+    {
+        return [];
+    }
+}
