+++ src/Elewant/Webapp/DomainModel/Herding/DesiredBreed.php
@@ -2,19 +2,26 @@
 
 declare(strict_types=1);
 
-namespace Elewant\AppBundle\Entity;
+namespace Elewant\Webapp\DomainModel\Herding;
 
+/**
+ * @todo Is it ok to use Herding\DomainModel here?
+ */
+
+use DateTime;
 use Doctrine\ORM\Mapping as ORM;
-use Elewant\Herding\Model\Breed;
+use Elewant\Herding\DomainModel\Breed\Breed;
 
 /**
  * @ORM\Entity
  * @ORM\Table(name="desired_breed", indexes={@ORM\Index(columns={"desired_on"})})
+ *
+ * This entity has a companion proxy, therefor is not final.
  */
 class DesiredBreed
 {
     /**
-     * @ORM\ManyToOne(targetEntity="Elewant\AppBundle\Entity\Herd", inversedBy="desiredBreeds")
+     * @ORM\ManyToOne(targetEntity="Herd", inversedBy="desiredBreeds")
      * @ORM\JoinColumn(name="herd_id", referencedColumnName="herd_id", nullable=false)
      * @ORM\Id
      * @var Herd
@@ -30,7 +37,7 @@ class DesiredBreed
 
     /**
      * @ORM\Column(type="datetime")
-     * @var \DateTime
+     * @var DateTime
      */
     private $desiredOn;
 
@@ -44,7 +51,7 @@ class DesiredBreed
         return $this->breed;
     }
 
-    public function desiredOn(): \DateTime
+    public function desiredOn(): DateTime
     {
         return $this->desiredOn;
     }
